# Multithreaded, selective attribute updates