--- a/src/GafferScene/InteractiveRender.cpp
+++ b/src/GafferScene/InteractiveRender.cpp
@@ -685,9 +685,13 @@ void InteractiveRender::runPipeline(tbb::pipeline *p)
 	
 	p->run( 2 * tbb::task_scheduler_init::default_num_threads() );
 }
-
+#include "sys/time.h"
 void InteractiveRender::outputScene( bool update )
 {
+	struct timeval tv;
+	gettimeofday(&tv, NULL);
+	double t0 = tv.tv_sec + double( tv.tv_usec )/ 1000000;
+	
 	SceneGraphIteratorFilter iterator( m_sceneGraph.get() );
 
 	SceneGraphEvaluatorFilter evaluator(
@@ -715,6 +719,11 @@ void InteractiveRender::outputScene( bool update )
 		continue;
 	}
 	pipelineThread.join();
+	
+	gettimeofday(&tv, NULL);
+	double t1 = tv.tv_sec + double( tv.tv_usec )/ 1000000;
+	
+	std::cerr << "outputScene " << t1 - t0 << std::endl;
 }
 
 void InteractiveRender::update()