Initial scene output also goes directly to the renderer, rather than via procedurals, so we can block the main thread