@johnhaddon 2015-01-15
Have you run the `GafferRenderManTest.InteractiveRenderManTest` unit tests? I get the following crash when I do :

```
testAddLight (GafferRenderManTest.InteractiveRenderManRenderTest.InteractiveRenderManRenderTest) ... 3DL ERROR S2074: cannot assign variable of type 'uniform vector' to parameter 'uniform point from' (shader: 'pointlight')
terminate called after throwing an instance of 'tbb::captured_exception'
  what():  Unable to find mapping for output path
```

I also notice that you've started using `type& name` and `type* name` here and there rather than following our established `type &name` and `type *name` convention. Not a biggy, but I would like to keep the codebase consistent so if you could stick with the convention that would be great.

I'll dive into a proper review once we know what's going on with the test failure...