@davidsminor 2015-01-15
Gah - yeah, I was just running GafferSceneTest wasn't I - what was I thinking. I'll take a look