@davidsminor 2015-01-15
eh? I wrote out all the "myLovelyPlane" images that were getting tested, and they all seemed fine. The 'eck?