@davidsminor 2015-01-16
Just got some new performance stats for the robot from this morning, for doing a big shader edit hitting 2600 objects. This is from 3delight 11.0.133, where they've sped up the RiEdit stuff quite a bit:
```
setup:                                          startup      edit
--------------------------------------------------------------------
3delight 11.0.133:                              23.6 sec     13 sec
no RiShader/Attribute calls:                    13 sec       12 sec
no RiEditBegin/End/Shader/Attribute calls:      13 sec       8 sec
all Ri calls commented out:                     13 sec       8 sec
gaffer traversal only:                          5 sec        4 sec

3delight 11.0.126:                              23.6 sec     30 sec
```
So 3delight 11.0.133 is considerably better than 3delight 11.0.126. It's interesting how commenting out the RiShader calls cuts the startup time in half.... it looks like most of the 3delight startup time is going into instantiating shaders? Anyway, it kind of looks like the IECoreRi code's costing us 8 seconds doesn't it? Traversal takes 4 seconds, but that happens in the background, so I reckon blasting through an equivalent precompiled list of IECoreRi calls in a single thread would probably take 8 seconds. On top of that, it looks like the RiEdit blocks cost us an extra 4 seconds