@johnhaddon 2015-01-19
Nice! 

Did you actually measure the `type == "shader"` change to `strcmp()` on its own? It sounds a bit fishy, because direct `string <-> const char *` comparisons are defined (and shouldn't make temporaries). If we really want to speed up the comparisons though, we should think about taking `InternedString` arguments in `Renderer` methods rather than `std::string`.

Now we have `GeometricTypedData`, we should be able to ditch the `typeHints` hack entirely. It would take a little bit of work in both IECoreRI and Gaffer to check that we were tracking the geometric type appropriately, but I think it'd be well worth doing - cleaner and faster.