@johnhaddon 2015-01-19
I've made a bunch of inline comments - a few style things that should be easy to fix, a couple of suggestions which should help with performance, and a couple of questions to aid my understanding. Thanks for the clear comments on the filter tasks - they made the code very easy to dive into.

There's one broader issue here as well though. It's not allowed to use "scene:path" values that aren't predetermined as being valid - where validity means "every name in the path is present in the childNames at that level" (SceneAlgo provides a check for this with `exists()`). Really, when we're doing the update traversal, we also need to be computing child names, and culling the traversal when we hit locations which are no longer valid. Do you think that's easy to fit in with the logic as it stands?