@davidsminor 2015-01-19
I guess we'd deal with the child names issue in plugDirtied()? Maybe store child names hashes in the SceneGraph, do a quick traversal if inPlug()->childNamesPlug() gets dirtied and zero out m_renderer/m_sceneGraph if we get any differences?