@davidsminor 2015-01-19
I guess we'd need to check if a location's scene:visible attribute changed too wouldn't we...