@johnhaddon 2015-01-20
I'm not 100% sure how we should deal with child names differences. I can imagine it being useful to switch between different graphs with subtly different child names, and keep the same IPR session going - for instance someone might switch between nodes where the only difference is that a light has been pruned. I think it might be best to continue with the IPR, but check child names just to avoid traversing into SceneGraph locations we know aren't valid for this traversal. Does that sound reasonable?