@davidsminor 2015-01-23
Hmm... I started putting something in the SceneGraphIteratorFilter that checked the child names whenever it was about to recurse into a new child. It was reading the childNames plug, then trying to std::find() the name of the child it was trying to go to in the result. It didn't seem to impact performance for the robot, but then I thought "quadratic scaling is fun!" and tried it on the snowy hillside. It looks like checking the child names/doing an std::find takes the startup time up from about 23 seconds to 60 seconds, and takes the shader edit time up from 16 seconds to 63 seconds. So that particular method is clearly a nono...

As a side note, it looks like my shader edit was taking 16 seconds largely because the 3delight render was still running while I was traversing the scene in parallel. If I split the pipeline into two phases, one where it pauses the render while traversing the scene, one where it emits the actual edits, the shader edit time goes down to 5 seconds (although this slows down the render startup quite a bit, and it means we have to compute the entire scene upfront before sending it to the renderer. This can be worked around, but it makes things a bit of a mess). It'd be really good if we could just pause the render while traversing the hierarchy/sending the edits wouldn't it - I've made a comment to that effect on that 3delight ticket about shader edits.